+++ coregrind/m_coredump/coredump-macho.c
@@ -36,7 +36,7 @@
 #include "pub_core_libcsetjmp.h"   // to keep _threadstate.h happy
 #include "pub_core_threadstate.h"
 
-void VG_(make_coredump)(ThreadId tid, const vki_siginfo_t *si, UInt max_size)
+void VG_(make_coredump)(ThreadId tid, const vki_siginfo_t *si, ULong max_size)
 {
    // DDD: #warning GrP fixme coredump
 }
